+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
+
+* Add `deleteFiles` option and `--delete-files` flag to delete all files/directories matching some glob patterns after conversion.
 <!-- Add new, unreleased changes here. -->
 
 ## [0.3.0] - 2017-11-28